Set CARGO env var to a valid cargo
Some build scripts read the CARGO env var and
use it to execute normal cargo scripts on it.
E.g. they expect ability to do "$CARGO run".
The cargo library sets $CARGO to our current
executable which resulted in unknown command
errors. This commit fixes this problem by
searching for a cargo in the path and setting
CARGO to it.