Use cargo dependency from git to allow people to use edition 2021
edition 2021 is not released yet, but some people already
try it out as beta and nightly compilers support it.

We use the git version of the cargo dependency to allow
these users to use cargo-udeps with their edition 2021
projects.

Generally we keep a policy of not using git dependencies in
cargo-udeps, but this time we make an exception due to the
significance of the 2021 edition. Once it releases, and
cargo 0.570 is published, we can use crates.io releases
again.